pyproject.toml
[project]
name = "kew-away-day"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.73.1",
    "python-dotenv>=1.2.2",
]
